requirements.txt
Django==4.0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
psycopg2-binary==2.9.4
celery==5.2.7
django-celery-beat==2.3.0
earthengine-api==0.1.329
gsconfig-py3==1.0.8

# drf-yasg[validation]==1.21.4
# drf-spectacular==0.24.2
# drf-spectacular[sidecar]

rasterio==1.3.4

#Test requirements
factory-boy==3.2.1
django-debug-toolbar==3.7.0